code agent/requirements.txt
fastapi
uvicorn
pytest
coverage
google-genai
python-multipart
